TRUNK: big commit containing: -sync with l2jdp till rev 2249 nearly finished(by decad) -sync with l2jc5 till rev 199(by evill33t) -DiezelMax's patch(implemented by decad) -fix for Zariche drop on death and hopefully fixed schuttgart respawn(by evill33t) -Clan hall implementation still has some debug lines, not finished in dp side and might have minor bugs, so CAUTION :D(by tomciaaa) -Finished clan skills(will be activated soon), fixed some bugs in clan halls and hopefully auctioneer... (by tomciaaa)
git-svn-id: http://svn.l2jfree.com/l2jfree@14 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/L2_GameServer_c5/java/net/sf/l2j/gameserver/model/actor/stat/PetStat.java
@@ -24,7 +24,7 @@ public class PetStat extends SummonStat
 
     // =========================================================
     // Method - Public
-    public boolean addExp(int value)
+    public boolean addExp(long value)
     {
         if (!super.addExp(value)) return false;
